Nearly there for the all documents rewrite. Having trouble figuring out how layouts should now work.
The problem is that we need to make the decision of whether or not to
parse special directories /^_/.test(dirName). Parsing would allow us to
watch those files, however can complicate the workflow as they all get
added to @database. Whereas if we don't parse them, then they are
entirely handled by plugins - which is probably desirable… It is a
tough call, and not sure how to proceed.