Fix #811
Turns out that resetCollections was setting docpad.generated to false.
But our docpad.generated = true assignment would only fire if
opts.initial was true. However, that would not be true if we are a
non-initial reset generation.

I’ve moved the reset task to be the first, and moved the opts
assignment after it. So that opts can be assigned based on their
correct values which can be determined after the reset tasks fires or
doesn’t fire.

I’ve also added a check for this invalid state in case it ever occurs
again.