Add colorful `inspect()` function to `docpadUtil`
Just a simple wrapper around Node’s built-in `util.inspect`.

Also, many calls to the built-in `util.inspect` are `require` -ed on each invocation.  It’s probably cached by Node, but it still feels a tad messy.

This way, we only `require` the built-in `util.inspect` function once, then call it when needed.